Create sql_test.sql
Write pseudo-SQL statements to create database tables to store the products of a basic webshop. Each product has a name, a price, a creation date and may belong to several categories. Categories have a name and a flag to indicate whether the category is private or public.
Write a SQL query to find the list of products that belong to more than 5 public categories.

Answer:
I created 3 tables: 
    - 1 table for products (contain a product's info)
    - 1 for categories (contains a category's info)
    - 1 for product and categories (contains the relation between the products and categories)